### Router * moved some component into their own folder * removed checkPermissions() and SYSADMIN from this package * Updated Reroute, SwitchWithError, and RouteDirector
