+++ packages/router/src/SwitchWithError/SwitchWithError.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import type {ChildrenArray} from 'react';
 import {Route, Switch} from 'react-router-dom';
-import NotFound from './components/NotFound';
+import NotFound from '../components/NotFound';
 
 type Props = {
 	children: ChildrenArray<*>,
