+++ packages/router/src/index.js
@@ -2,13 +2,10 @@ import Reroute from './Reroute';
 import SwitchWithError from './SwitchWithError';
 import RouteDirector from './RouteDirector';
 import lazy from './lazy';
-import checkPermissions, {SYSADMIN} from './checkPermissions';
 
 export {
 	Reroute,
 	SwitchWithError,
 	RouteDirector,
 	lazy,
-	checkPermissions,
-	SYSADMIN,
 };
