--- packages/router/src/index.server.js
@@ -1,6 +1,0 @@
-import checkPermissions, {SYSADMIN} from './checkPermissions';
-
-export {
-	checkPermissions,
-	SYSADMIN,
-};
